+++ src/main/java/io/github/notenoughmail/kubejstfc/compat/worldjs/builders/forest/KrummholzBuilder.java
@@ -31,7 +31,7 @@ public class KrummholzBuilder extends ConfiguredFeatureBuilder.WithFeature<Krumm
     @Info("The krummholz block to place")
     public KrummholzBuilder krummholz(Block k) {
         final Collection<Property<?>> props = k.getStateDefinition().getProperties();
-        if (!(props.contains(KrummholzBlock.TIP) || props.contains(KrummholzBlock.BOTTOM))) {
+        if (!props.contains(KrummholzBlock.TIP) || !props.contains(KrummholzBlock.BOTTOM)) {
             throw exception("Krummholz block must have tip and bottom block properties!")
                     .customData("tip property", KrummholzBlock.TIP)
                     .customData("bottom property", KrummholzBlock.BOTTOM)